Updated the functionalities

+++ Empower-Quiz-application/src/main/java/com/example/quizapplication/Service/QuestionService.java
@@ -1,10 +1,13 @@
 package com.example.quizapplication.Service;
 
 import com.example.quizapplication.DAO.QuestionDao;
-import com.example.quizapplication.Question;
+import com.example.quizapplication.model.Question;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -12,21 +15,50 @@ public class QuestionService {
 
     @Autowired
     QuestionDao questionDao;
-    public List<Question> getallQuesions() {
-        return questionDao.findAll();
+    public ResponseEntity<List<Question>> getallQuesions() {
+        try {
+            return new ResponseEntity<>(questionDao.findAll(), HttpStatus.OK);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return new ResponseEntity<>(new ArrayList<>(), HttpStatus.BAD_REQUEST);
     }
 
-    public List<Question> getQuestionsByCategory(String category) {
-         return questionDao.findByCategory(category);
+    public ResponseEntity<List<Question>> getQuestionsByCategory(String category) {
+        try {
+            return new ResponseEntity<>(questionDao.findByCategory(category), HttpStatus.OK);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return new ResponseEntity<>(new ArrayList<>(), HttpStatus.BAD_REQUEST);
     }
 
-    public String addQuestion(Question question) {
-        questionDao.save(question);
-        return "Success";
+    public ResponseEntity<String> addQuestion(Question question) {
+        try {
+            questionDao.save(question);
+            return new ResponseEntity<>("Success", HttpStatus.CREATED);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return new ResponseEntity<>("Failed", HttpStatus.BAD_REQUEST);
     }
 
-    public String deleteQuestion(Integer id) {
-         questionDao.deleteById(id);
-         return "deleted Successfully";
-    }
-}
+
+    public ResponseEntity<String> deleteQuestion(Integer id) {
+
+        try{
+
+            if (!questionDao.existsById(id)) {
+                return new ResponseEntity<>("Question not found", HttpStatus.NOT_FOUND);
+            }
+
+            questionDao.deleteById(id);
+            return new ResponseEntity<>("Deleted successfully", HttpStatus.OK);
+        }
+        catch(Exception e){
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("Failed to delete question");
+        }
+
+}}
